+++ packages/components/CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## [Unreleased]
 
+### Changed
+
+- **Stylesheets**: Bundle Porsche Next fonts with the stylesheets for self-hosting, instead of loading fonts from a CDN.
+
 ## [4.6.0] - 2026-08-20
 
 ## [4.6.0-rc.3] - 2026-08-19